requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-BasicAuth==0.2.0
idna==2.6
influxdb==5.0.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
paho-mqtt==1.3.1
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
